Move some code to separate files
